// src/types/api/educationItem.ts
export interface EducationItem {
  EducationItemID?: number;
  EducationListID?: number;
  PersonID?: number;
  InstitutionName?: string;
  EducationDescription?: string;
  CompletionDate?: string;
  CompletionDateString?: string;
  CreationDate?: Date;
  CreatedBy?: number;
  LastModifiedDate?: Date;
  LastModifiedBy?: number;
}